Bug fixes and variants (#71)
* Change EvalEngine so it only runs when Eval Engine panel is visible

* Fix bugs when logging in as an unregistered user

* Split up multi-line messages properly. I.e. FICS uses 'fics%' to separate different multi-line messages when sent together.
'iset defprompt 1' forces 'fics%' to be used even if user changes their prompt.

* Fix bug games list not being parsed properly

* Add buffered forward/back buttons (which can be clicked multiple times quickly before server responds to first forward click etc)

* Simplify board flipping code

* Fix chat tab bugs
Don't mark tab as 'unviewed' when new message is from ourselves (this.user)
Prefix tab id with 'tab-' to stop code breaking when someone is named the same as an id in our html (e.g. puzzlebot).
Display 'Sorry the message is too long.' as chat notification instead of to console
Still send error messages as notifications when Notifications turned off in settings

* Add support for variants: Fischer Random, other Wild variants and Losers
Add support for black starting in the initial position (i.e. for puzzlebot)

* Fix bug where space characters were being removed from chat messages (by FICS server) due to wrapping

---------

Co-authored-by: Free Chess Club Dev <hi@freechess.club>