Another bug fix
When engine analysis is on, and you sought a new game without actually stopping the engine, the autoshape (yellow arrow) would still be visible on the board and there was no way to make it go away which was somewhat distracting.

My earlier fix of "board.setAutoShapes([])" when stopping engine didn't actually work (maybe a chessground bug, not sure?). So this fix just explicitly sets autoShapes to empty when initializing/updating the board.